requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.5.0
tiktoken>=0.5.2
streamlit>=1.34.0
pyyaml>=6.0.2
